# api/openapi.yaml
openapi: 3.0.3
info:
  title: HopeItWorks API
  description: |
    Backend API for the HopeItWorks project management platform.
    This is the single source of truth for the API contract.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: HopeItWorks Team
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - cookieAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List all projects
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - $ref: "#/components/parameters/SortByParam"
      responses:
        "200":
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createProject
      summary: Create a new project
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /projects/{id}:
    get:
      operationId: getProject
      summary: Get a project by ID
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateProject
      summary: Update a project
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteProject
      summary: Delete a project
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Epics ──────────────────────────────────────────────────────────────
...
  /projects/{projectId}/costs:
    get:
      operationId: getProjectCosts
      summary: Get aggregated cost data for a project
      description: >
        Returns aggregated cost data for the specified project over a time period,
        with breakdowns by story, run, and model. Budget limits are included for
        informational display alongside usage.
      tags: [costs]
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
          description: Time period for aggregation
      responses:
        "200":
          description: Aggregated project cost data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCostSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/stories/{storyId}/costs:
    get:
      operationId: getStoryCosts
      summary: Get total costs for a story across all runs
      tags: [costs]
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/StoryIdPath"
      responses:
        "200":
          description: Story cost summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryCostSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/runs/{runId}/costs:
    get:
      operationId: getRunCosts
      summary: Get total costs for a run with per-step breakdown
      tags: [costs]
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/RunIdPath"
      responses:
        "200":
          description: Run cost detail with per-step breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCostDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

# ── Components ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token sent as an httpOnly secure cookie
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Project UUID
...
    RunIdPath:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Run UUID
    StoryIdPath:
      name: storyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Story UUID
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    SortByParam:
      name: sort_by
      in: query
      required: false
      schema:
        type: string
        default: created_at
      description: Field to sort by
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - owner_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: My Project
        description:
          type: string
          example: A sample project
        owner_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        max_budget:
          type: number
          format: double
          nullable: true
          description: Maximum budget for the project in USD (informational only)
          example: 100.00
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: My Project
        description:
          type: string
          maxLength: 1000
          example: A sample project
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Updated Project Name
        description:
          type: string
          maxLength: 1000
          example: Updated description
        max_budget:
          type: number
          format: double
          nullable: true
          description: Maximum budget for the project in USD (informational only)
          example: 100.00
    ProjectList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ── Epic ──────────────────────────────────────────────────────────
...
    ProjectCostSummary:
      type: object
      required:
        - total_cost
        - total_input
        - total_output
        - by_story
        - by_run
        - by_model
      properties:
        total_cost:
          type: number
          format: double
          description: Total cost in USD for the period
          example: 12.50
        total_input:
          type: integer
          format: int64
          description: Total input tokens for the period
          example: 500000
        total_output:
          type: integer
          format: int64
          description: Total output tokens for the period
          example: 100000
        max_budget:
          type: number
          format: double
          nullable: true
          description: Project budget limit in USD (informational only)
          example: 100.00
        by_story:
          type: array
          items:
            $ref: "#/components/schemas/StoryCostBreakdown"
        by_run:
          type: array
          items:
            $ref: "#/components/schemas/RunCostBreakdown"
        by_model:
          type: array
          items:
            $ref: "#/components/schemas/ModelCostBreakdown"
    StoryCostBreakdown:
      type: object
      required:
        - story_id
        - story_key
        - total_cost
      properties:
        story_id:
          type: string
          format: uuid
        story_key:
          type: string
          example: S-01
        total_cost:
          type: number
          format: double
          example: 5.25
    RunCostBreakdown:
      type: object
      required:
        - run_id
        - story_key
        - status
        - total_cost
        - created_at
      properties:
        run_id:
          type: string
          format: uuid
        story_key:
          type: string
          example: S-01
        status:
          type: string
          example: completed
        total_cost:
          type: number
          format: double
          example: 3.15
        created_at:
          type: string
          format: date-time
    ModelCostBreakdown:
      type: object
      required:
        - model
        - total_cost
        - tokens_input
        - tokens_output
      properties:
        model:
          type: string
          example: claude-opus-4-6
        total_cost:
          type: number
          format: double
          example: 10.00
        tokens_input:
          type: integer
          format: int64
          example: 300000
        tokens_output:
          type: integer
          format: int64
          example: 50000
    StoryCostSummary:
      type: object
      required:
        - story_id
        - total_cost
        - total_input
        - total_output
        - run_count
      properties:
        story_id:
          type: string
          format: uuid
        total_cost:
          type: number
          format: double
          example: 5.25
        total_input:
          type: integer
          format: int64
          example: 200000
        total_output:
          type: integer
          format: int64
          example: 40000
        run_count:
          type: integer
          description: Number of distinct runs that incurred costs
          example: 3
    RunCostDetail:
      type: object
      required:
        - run_id
        - total_cost
        - steps
      properties:
        run_id:
          type: string
          format: uuid
        total_cost:
          type: number
          format: double
          example: 3.15
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepCostBreakdown"
    StepCostBreakdown:
      type: object
      required:
        - step_id
        - step_name
        - model
        - tokens_input
        - tokens_output
        - cost_usd
      properties:
        step_id:
          type: string
          format: uuid
        step_name:
          type: string
          example: implement
        model:
          type: string
          example: claude-opus-4-6
        tokens_input:
          type: integer
          format: int64
          example: 100000
        tokens_output:
          type: integer
          format: int64
          example: 20000
        cost_usd:
          type: number
          format: double
          example: 3.00

    # ── Shared ─────────────────────────────────────────────────────────
    Pagination:
      type: object
      required:
        - total
        - page
        - per_page
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: USER_NOT_FOUND
            message:
              type: string
              example: "User with ID abc123 not found"
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...