# backend/internal/adapter/postgres/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package postgres

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type CostRecord struct {
	ID           uuid.UUID      `json:"id"`
	RunStepID    uuid.UUID      `json:"run_step_id"`
	ProjectID    uuid.UUID      `json:"project_id"`
	TokensInput  int64          `json:"tokens_input"`
	TokensOutput int64          `json:"tokens_output"`
	CostUsd      pgtype.Numeric `json:"cost_usd"`
	Model        string         `json:"model"`
	CreatedAt    time.Time      `json:"created_at"`
}

type Epic struct {
	ID          uuid.UUID   `json:"id"`
	ProjectID   uuid.UUID   `json:"project_id"`
	Name        string      `json:"name"`
	Description pgtype.Text `json:"description"`
	Status      string      `json:"status"`
	CreatedAt   time.Time   `json:"created_at"`
	UpdatedAt   time.Time   `json:"updated_at"`
}

type Event struct {
	ID         uuid.UUID `json:"id"`
	ProjectID  uuid.UUID `json:"project_id"`
	EntityType string    `json:"entity_type"`
	EntityID   uuid.UUID `json:"entity_id"`
	Action     string    `json:"action"`
	Payload    []byte    `json:"payload"`
	CreatedAt  time.Time `json:"created_at"`
}

type PipelineConfig struct {
	ID         uuid.UUID `json:"id"`
	ProjectID  uuid.UUID `json:"project_id"`
	ConfigYaml string    `json:"config_yaml"`
	Version    int32     `json:"version"`
	CreatedAt  time.Time `json:"created_at"`
	UpdatedAt  time.Time `json:"updated_at"`
}

type Project struct {
	ID           uuid.UUID      `json:"id"`
	Name         string         `json:"name"`
	Description  pgtype.Text    `json:"description"`
	OwnerID      pgtype.UUID    `json:"owner_id"`
	RepoUrl      pgtype.Text    `json:"repo_url"`
	GitProvider  string         `json:"git_provider"`
	GitTokenEnv  pgtype.Text    `json:"git_token_env"`
	AgentRuntime string         `json:"agent_runtime"`
	DefaultModel pgtype.Text    `json:"default_model"`
	MaxBudget    pgtype.Numeric `json:"max_budget"`
	CreatedAt    time.Time      `json:"created_at"`
	UpdatedAt    time.Time      `json:"updated_at"`
}

type ProjectUser struct {
	ProjectID uuid.UUID `json:"project_id"`
	UserID    uuid.UUID `json:"user_id"`
	Role      string    `json:"role"`
	CreatedAt time.Time `json:"created_at"`
}

type PromptTemplate struct {
	ID              uuid.UUID `json:"id"`
	ProjectID       uuid.UUID `json:"project_id"`
	Name            string    `json:"name"`
	TemplateContent string    `json:"template_content"`
	Type            string    `json:"type"`
	CreatedAt       time.Time `json:"created_at"`
	UpdatedAt       time.Time `json:"updated_at"`
}

type Run struct {
	ID                     uuid.UUID          `json:"id"`
	ProjectID              uuid.UUID          `json:"project_id"`
	StoryID                uuid.UUID          `json:"story_id"`
	Status                 string             `json:"status"`
	PipelineConfigSnapshot []byte             `json:"pipeline_config_snapshot"`
	StartedAt              pgtype.Timestamptz `json:"started_at"`
	CompletedAt            pgtype.Timestamptz `json:"completed_at"`
	ErrorMessage           pgtype.Text        `json:"error_message"`
	CreatedAt              time.Time          `json:"created_at"`
	UpdatedAt              time.Time          `json:"updated_at"`
}

type RunStep struct {
	ID           uuid.UUID          `json:"id"`
	RunID        uuid.UUID          `json:"run_id"`
	StepName     string             `json:"step_name"`
	StepOrder    int32              `json:"step_order"`
	Action       string             `json:"action"`
	Status       string             `json:"status"`
	StartedAt    pgtype.Timestamptz `json:"started_at"`
	CompletedAt  pgtype.Timestamptz `json:"completed_at"`
	ErrorMessage pgtype.Text        `json:"error_message"`
	ContainerID  pgtype.Text        `json:"container_id"`
	LogTail      pgtype.Text        `json:"log_tail"`
	CreatedAt    time.Time          `json:"created_at"`
}

type Story struct {
	ID                 uuid.UUID   `json:"id"`
	ProjectID          uuid.UUID   `json:"project_id"`
	EpicID             pgtype.UUID `json:"epic_id"`
	Key                string      `json:"key"`
	Title              string      `json:"title"`
	Objective          pgtype.Text `json:"objective"`
	TargetFiles        []byte      `json:"target_files"`
	DependsOn          []byte      `json:"depends_on"`
	Scope              pgtype.Text `json:"scope"`
	Status             string      `json:"status"`
	AcceptanceCriteria pgtype.Text `json:"acceptance_criteria"`
	CreatedAt          time.Time   `json:"created_at"`
	UpdatedAt          time.Time   `json:"updated_at"`
}

type User struct {
	ID           uuid.UUID          `json:"id"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"password_hash"`
	Name         string             `json:"name"`
	Role         string             `json:"role"`
	CreatedAt    time.Time          `json:"created_at"`
	UpdatedAt    time.Time          `json:"updated_at"`
	DeletedAt    pgtype.Timestamptz `json:"deleted_at"`
}